+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Invoice resource to load your account with dynamic QR Codes
+
 ## [2.1.0] - 2020-10-28
 ### Added
 - BoletoHolmes to investigate boleto status according to CIP